refactor: SwapChecker into component
removes some createRoot weirdness and makes tests also easier
