+++ src/index.tsx
@@ -3,12 +3,13 @@ import "@fontsource/noto-sans";
 import "@fontsource/noto-sans/800.css";
 import { Navigate, Route, Router, Routes } from "@solidjs/router";
 import log from "loglevel";
-import { Show, createRoot, createSignal } from "solid-js";
+import { Show, createSignal } from "solid-js";
 import { render } from "solid-js/web";
 
 import Footer from "./components/Footer";
 import Nav from "./components/Nav";
 import Notification from "./components/Notification";
+import { SwapChecker } from "./components/SwapChecker";
 import { loglevel, network } from "./config";
 import { Web3SignerProvider } from "./context/Web3";
 import { detectLanguage } from "./i18n/detect";
@@ -25,7 +26,6 @@ import "./style/index.scss";
 import { detectEmbedded } from "./utils/embed";
 import { checkReferralId } from "./utils/helper";
 import "./utils/patches";
-import { swapChecker } from "./utils/swapChecker";
 import { checkWasmSupported } from "./utils/wasmSupport";
 import { detectWebLNProvider } from "./utils/webln";
 
@@ -37,10 +37,6 @@ detectWebLNProvider().then((state: boolean) => setWebln(state));
 setWasmSupported(checkWasmSupported());
 checkReferralId();
 detectLanguage();
-
-createRoot(() => {
-    swapChecker();
-});
 
 if (detectEmbedded()) {
     setHideHero(true);
@@ -61,6 +57,7 @@ const cleanup = render(
     () => (
         <Router>
             <Web3SignerProvider>
+                <SwapChecker />
                 <Show when={!embedded()}>
                     <Nav network={network} />
                 </Show>
